rename svg folder
